Movie Recommendation/requirements.txt
pandas==1.3.3
scikit-surprise==1.1.1
tqdm==4.62.3
joblib==1.1.0
